Fixed bugs.
